pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "postall"
version = "3.0.0"
description = "PostAll - Automated Social Media Content Publishing System"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "PostAll Contributors"}
]
keywords = [
    "social-media",
    "content-marketing",
    "automation",
    "ai",
    "content-generation"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Marketing",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "pillow>=9.0.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
cloud = [
    "aiohttp>=3.8.0",
    "aiosqlite>=0.19.0",
]
all = [
    "postall[dev,cloud]",
    "tweepy>=4.14.0",
    "openai>=1.0.0",
]

[project.scripts]
postall = "postall.cli:main"

[project.urls]
Homepage = "https://github.com/postall/postall"
Documentation = "https://github.com/postall/postall#readme"
Repository = "https://github.com/postall/postall"
Issues = "https://github.com/postall/postall/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["postall*"]

[tool.setuptools.package-data]
postall = [
    "prompts/*.md",
    "config/*.yaml",
]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=postall"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
